+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Handle not a number errors on `!spellbot awards` command.
+
 ## [v6.1.2](https://github.com/lexicalunit/spellbot/releases/tag/v6.1.2) - 2021-06-01
 
 ### Changed